hw 18 blogs sa fix2

+++ src/features/sa/use-cases/deletePostsSaBlog.ts
@@ -1,12 +1,13 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { BlogsSaService } from '../api/blogs.sa.service';
 import { PostsSaService } from '../api/posts.sa.service';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 
 export class DeletePostsSaBlog {
   constructor(
     public readonly blogId: string,
     public readonly postId: string,
+    public readonly userId: string,
   ) {}
 }
 
@@ -21,8 +22,10 @@ export class DeletePostsSaBlogHandler
   async execute(command: DeletePostsSaBlog) {
     const findBlog = await this.blogsSaService.findBlogId(command.blogId);
     if (!findBlog) throw new NotFoundException();
+    if (findBlog.userId !== command.userId) throw new ForbiddenException();
     const findPost = await this.postsSaService.findPostId(command.postId);
     if (!findPost) throw new NotFoundException();
-    return await this.postsSaService.deletePostsForBlogId(command.postId);
+    if (findPost.blogId !== command.blogId)
+      return await this.postsSaService.deletePostsForBlogId(command.postId);
   }
 }
